fix: UIディレクトリを削除してコンポーネントを整理
